+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/commonpage/CommonModel.java
@@ -1,4 +1,4 @@
-package com.qatarmuseums.qatarmuseumsapp.commonactivity;
+package com.qatarmuseums.qatarmuseumsapp.commonpage;
 
 
 public class CommonModel {
