+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/education/EducationAdapter.java
@@ -1,4 +1,4 @@
-package com.qatarmuseums.qatarmuseumsapp.educationactivity;
+package com.qatarmuseums.qatarmuseumsapp.education;
 
 import android.app.Dialog;
 import android.content.Context;
