+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/home/HomeList.java
@@ -1,4 +1,4 @@
-package com.qatarmuseums.qatarmuseumsapp.homeactivity;
+package com.qatarmuseums.qatarmuseumsapp.home;
 
 
 public class HomeList {
