+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/museum/MuseumActivity.java
@@ -1,22 +1,14 @@
 package com.qatarmuseums.qatarmuseumsapp.museum;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SnapHelper;
-import android.view.Gravity;
-import android.view.MotionEvent;
 import android.view.View;
-import android.widget.HorizontalScrollView;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
 
-import com.github.rubensousa.gravitysnaphelper.GravitySnapHelper;
 import com.qatarmuseums.qatarmuseumsapp.R;
-import com.qatarmuseums.qatarmuseumsapp.baseactivity.BaseActivity;
+import com.qatarmuseums.qatarmuseumsapp.base.BaseActivity;
 
 import java.util.ArrayList;
 import java.util.List;
