+++ app/src/main/res/layout/activity_base.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.baseactivity.BaseActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.base.BaseActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
