+++ app/src/main/res/layout/activity_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.calendaractivity.CalendarActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.calendar.CalendarActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
