+++ app/src/main/res/layout/activity_dining.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.detailsactivity.DetailsActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.detailspage.DetailsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
