+++ app/src/main/res/layout/activity_education_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.educationactivity.EducationCalendarActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.education.EducationCalendarActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
