+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.homeactivity.HomeActivity">
+    tools:context="com.qatarmuseums.qatarmuseumsapp.home.HomeActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recycler_view"
