+++ app/src/main/res/layout/content_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.baseactivity.BaseActivity"
+    tools:context="com.qatarmuseums.qatarmuseumsapp.base.BaseActivity"
     tools:showIn="@layout/activity_base">
