+++ app/src/main/res/layout/dining_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.qatarmuseums.qatarmuseumsapp.detailsactivity.DetailsActivity"
+    tools:context="com.qatarmuseums.qatarmuseumsapp.detailspage.DetailsActivity"
     tools:showIn="@layout/activity_details">
 
     <LinearLayout
